decouple audio size from grid size
add the ability to set the length of saved audio frames independent of
the number of rendered grid squares. exports this as a parameter in the
audio settings. updated the render shader to use textures so that the
width is not fixed. also found some bug so had to rewrite the update
pass using the new Program API.